### 2021-01-28 #2994 @whedon
[PRE REVIEW]: PyFstat: a Python package for continuous gravitational-wave data analysis
**Submitting author:** @dbkeitel (<a href="http://orcid.org/0000-0002-2824-626X">David Keitel</a>)
**Repository:** <a href="https://github.com/PyFstat/PyFstat/" target ="_blank">https://github.com/PyFstat/PyFstat/</a>
**Version:** v1.11.1
