### 2021-01-28 #2994 @arfon
I think the presence of a `paper.tex` file is confusing Whedon: https://github.com/PyFstat/PyFstat/blob/joss/joss-paper/paper.tex

I would suggest removing this and trying again.
