### 2021-01-28 #2994 @danielskatz
👋 @dbkeitel  - can you remove paper.tex from the repo, and then try the compilation, by putting ```@whedon generate pdf from branch joss``` in a new comment here?  I'm going to mark this as paused for the minute
