### 2021-01-28 80ad938
Delete paper.tex
