### 2021-01-28 #2994 @dbkeitel
If those first wo people you asked aren't available, cmbiwer would be a great choice (GW analyst with experience developing similar packages, but no direct contact with this project).

Some more likely-looking people (with python & astro listed as language/field) going up from the bottom of the table would be e.g.: RobertRosca arjunsavel sibirrer JBorrow cavestruz .
