### 2021-01-28 #2994 @khanx169
> 👋 @khanx169, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html (You were suggested by @eliu23)

@danielskatz yes I'd be happy to review.
