### 2021-02-11 0b0f096
pin numpy<1.20 to fix lalsuite incompatibility
 - see https://git.ligo.org/lscsoft/lalsuite/-/issues/414
