+++ setup.py
@@ -54,7 +54,7 @@ setup(
     ],
     python_requires=">=%s.%s.%s" % min_python_version[:3],
     install_requires=[
-        "numpy",
+        "numpy<1.20",
         "matplotlib>=2.1",
         "scipy",
         "ptemcee",
