### 2021-02-11 #3000 @dbkeitel
Thanks from my side too! There's apparently a numpy-lalsuite version incompatibility at the moment that breaks our automated tests, I'm looking into fixing it by a temporary version pinning.
