### 2021-02-11 618bb88
update citation requests in README
