### 2021-02-11 f3130d7
update changelog for v1.11.2
