### 2021-02-11 34ca4f2
Juggle input params to circumvent a parent check
- Checks in `_basic_setup` clash against different options in Artifact.
- Now it works.
