### 2021-02-11 5d52612
Update pyfstat/make_sfts.py
Co-authored-by: David Keitel <david.keitel@ligo.org>
