### 2021-02-11 8e1d1e5
Start implementing TransientLineWriter
- Clean up some (but not all!) v4 options

- Add the whole set of transient window to writer

- Add signal parameters to v4

Adadpt tests to deal with the obligatory IFO

Adapt input values (F0, Band) to MFD_v4's requirements
