### 2021-02-11 af6d263
fix and test FrequencyModulatedArtifactWriter
 - needed to adapt concatenate_sft_files()
   method to changed lalapps_splitSFTs filename convention
