+++ pyfstat/make_sfts.py
@@ -1378,39 +1378,40 @@ class FrequencyModulatedArtifactWriter(Writer):
         return self.h0
 
     def concatenate_sft_files(self):
-        """Deals with multiple SFT files via lalapps_splitSFTs executable."""
-        SFTFilename = lalpulsar.OfficialSFTFilename(
-            self.detectors[0],
-            self.detectors[1],
-            self.nsfts,
-            self.Tsft,
-            int(self.tstart),
-            int(self.duration),
-            self.label,
+        """Merges the individual SFT files via lalapps_splitSFTs executable."""
+
+        SFTFilename = (
+            f"{self.detectors[0]}-{self.nsfts}_{self.detectors}_{self.Tsft}SFT"
         )
+        # We don't try to reproduce the NB filename convention exactly,
+        # as there could be always rounding offsets with the number of bins,
+        # instead we use wildcards there.
+        outfreq = int(np.floor(self.fmin))
+        outwidth = int(np.floor(self.Band))
+        SFTFilename += f"_NB_F{outfreq:04d}Hz*_W{outwidth:04d}Hz*"
+        SFTFilename += f"-{self.tstart}-{self.duration}.sft"
         SFTFile_fullpath = os.path.join(self.outdir, SFTFilename)
 
-        # If the file already exists, simply remove it for now (no caching
-        # implemented)
-        helper_functions.run_commandline(
-            "rm {}".format(SFTFile_fullpath), raise_error=False, log_level=10
-        )
+        if os.path.isfile(SFTFile_fullpath):
+            logging.info(
+                f"Removing previous file(s) {SFTFile_fullpath} (no caching implemented)."
+            )
+            helper_functions.run_commandline(f"rm {SFTFile_fullpath}")
 
         inpattern = os.path.join(self.tmp_outdir, "*sft")
-        cl_splitSFTS = "lalapps_splitSFTs -fs {} -fb {} -fe {} -o {} -i {}".format(
-            self.fmin, self.Band, self.fmin + self.Band, SFTFile_fullpath, inpattern
+        cl_splitSFTS = "lalapps_splitSFTs -fs {} -fb {} -fe {} -n {} -- {}".format(
+            self.fmin, self.Band, self.fmin + self.Band, self.outdir, inpattern
         )
         helper_functions.run_commandline(cl_splitSFTS)
-        helper_functions.run_commandline("rm {} -r".format(self.tmp_outdir))
-        files = glob.glob(SFTFile_fullpath + "*")
-        if len(files) == 1:
-            fn = files[0]
-            fn_new = fn.split(".")[0] + ".sft"
-            helper_functions.run_commandline("mv {} {}".format(fn, fn_new))
-        else:
+        helper_functions.run_commandline(f"rm -r {self.tmp_outdir}")
+        outglob = glob.glob(SFTFile_fullpath)
+        if len(outglob) != 1:
             raise IOError(
-                "Attempted to rename file, but multiple files found: {}".format(files)
+                f"Expected to produce exactly 1 merged file matching pattern '{SFTFile_fullpath}',"
+                f" but got {len(outglob)} matches: {outglob}. Something went wrong!"
             )
+        self.sftfilepath = outglob[0]
+        logging.info(f"Successfully wrote SFTs to: {self.sftfilepath}")
 
     def pre_compute_evolution(self):
         """Precomputes evolution parameters for the artifact.
@@ -1458,7 +1459,6 @@ class FrequencyModulatedArtifactWriter(Writer):
         This loops over SFTs and generate them serially or in parallel,
         then contatenates the results together at the end.
         """
[…]
