+++ CHANGELOG.md
@@ -1,8 +1,8 @@
-## 1.11.2 [11/02/2021]
+## 1.11.2 [12/02/2021]
 
  - pinned numpy dependency to <1.20 to fix incompatibility with lalsuite 6.81
  - updated ephemerides instructions and citation requests in README
- - fixed setup of FrequencyModulatedArtifactWriter class
+ - fixed FrequencyModulatedArtifactWriter class
 
 ## 1.11.1 [26/01/2021]
 
