### 2021-02-11 5f47bea
avoid duplicate checks on internal PRs
- solution by dantup & johnbillion at https://github.community/t/duplicate-checks-on-push-and-pull-request-simultaneous-event/18012
