### 2021-02-12 d1443dd
FrequencyModulatedArtifactWriter : check that detectors is a single-IFO string
