### 2021-02-12 31b53c9
Implement line writer
