### 2021-02-12 c710082
Add options to re-scale amplitude
