### 2021-02-12 3ab403c
remove tests from compiled docs
