+++ docs/source/tests.rst
@@ -1,7 +1,0 @@
-tests module
-============
-
-.. automodule:: tests
-   :members:
-   :undoc-members:
-   :show-inheritance:
