### 2021-02-15 71cd22a
Remove unnecessary dependence on sphinx
- Consequence of #282
