### 2021-02-15 8ce06f5
Ammend comments and some docstrings
- This must be finished once a common nomenclature is decided
