### 2021-02-15 6f02e73
We decided to use LineWriter
- Change names accordingly
