### 2021-02-15 dba5391
Fix documentation of fminBand
