### 2021-02-15 9adadcb
Change h0_eff implementation
