### 2021-02-15 4cb7c2c
Fix docstring typos
