### 2021-02-15 fde13d6
Apply  suggestions from code review
Co-authored-by: David Keitel <david.keitel@ligo.org>
