### 2021-02-15 b47658e
Implement line writer (#281)
