+++ pyfstat/make_sfts.py
@@ -191,6 +191,9 @@ class Writer(BaseSearchClass):
     for more detailed help with some of the parameters.
     """
 
+    mfd = "lalapps_Makefakedata_v5"
+    """The executable; can be overridden by child classes."""
+
     signal_parameter_labels = [
         "tref",
         "F0",
@@ -203,12 +206,16 @@ class Writer(BaseSearchClass):
         "psi",
         "phi",
         "transientWindowType",
+        "transientStartTime",
+        "transientTau",
     ]
     """Default convention of labels for the various signal parameters."""
 
     gps_time_and_string_formats_as_LAL = {
         "refTime": ":10.9f",
         "transientWindowType": ":s",
+        "transientStartTime": ":10.0f",
+        "transientTau": ":10.0f",
     }
     """Dictionary to ensure proper format handling for some special parameters.
 
@@ -509,12 +516,6 @@ class Writer(BaseSearchClass):
             for key in self.signal_parameters
         }
 
-        if self.signal_parameters["transientWindowType"] != "none":
-            self.signal_parameters["transientStartTime"] = self.transientStartTime
-            self.signal_formats["transientStartTime"] = ":10.0f"
-            self.signal_parameters["transientTau"] = self.transientTau
-            self.signal_formats["transientTau"] = ":10.0f"
-
     def calculate_fmin_Band(self):
         """Set fmin and Band for the output SFTs to cover.
 
@@ -749,9 +750,23 @@ class Writer(BaseSearchClass):
         to see if equivalent data files already exist,
         and else runs the actual generation code.
         """
+        cl_mfd = self._build_MFD_command_line()
 
-        mfd = "lalapps_Makefakedata_v5"
-        cl_mfd = [mfd]
+        check_ok = self.check_cached_data_okay_to_use(cl_mfd)
+        if check_ok is False:
+            helper_functions.run_commandline(cl_mfd)
+            if not np.all([os.path.isfile(f) for f in self.sftfilenames]):
+                raise IOError(
+                    f"It seems we successfully ran {self.mfd},"
+                    f" but did not get the expected SFT file path(s): {self.sftfilepath}."
+                )
+            logging.info(f"Successfully wrote SFTs to: {self.sftfilepath}")
+            logging.info("Now validating each SFT file...")
+            for sft in self.sftfilenames:
+                lalpulsar.ValidateSFTFile(sft)
+
+    def _build_MFD_command_line(self):
+        cl_mfd = [self.mfd]
         cl_mfd.append("--outSingleSFT=TRUE")
         cl_mfd.append('--outSFTdir="{}"'.format(self.outdir))
         cl_mfd.append('--outLabel="{}"'.format(self.label))
@@ -804,19 +819,7 @@ class Writer(BaseSearchClass):
         if self.randSeed:
             cl_mfd.append("--randSeed={}".format(self.randSeed))
 
-        cl_mfd = " ".join(cl_mfd)
-        check_ok = self.check_cached_data_okay_to_use(cl_mfd)
-        if check_ok is False:
-            helper_functions.run_commandline(cl_mfd)
-            if not np.all([os.path.isfile(f) for f in self.sftfilenames]):
-                raise IOError(
-                    f"It seems we successfully ran {mfd},"
-                    f" but did not get the expected SFT file path(s): {self.sftfilepath}."
-                )
-            logging.info(f"Successfully wrote SFTs to: {self.sftfilepath}")
-            logging.info("Now validating each SFT file...")
-            for sft in self.sftfilenames:
-                lalpulsar.ValidateSFTFile(sft)
+        return " ".join(cl_mfd)
 
     def predict_fstat(self, assumeSqrtSX=None):
         """Predict the expected F-statistic value for the injection parameters.
[…]
