### 2021-02-16 7a395e9
fix conda badge in README.md
