### 2021-02-22 be991ef
Remove h0_eff from the class
- What gets in, gets out. No black magic whatsoever.
