### 2021-02-22 e278ae3
Writer/Refactor required signal params
- This will allow a simpler inheritance to LineWriter
