### 2021-02-22 2fb4414
LineWrite/Improve inheritance
