### 2021-02-22 ada9212
integration.yml: add flake8-docstrings and flake8-executable
