### 2021-02-22 6636bd3
remove shebangs from example scripts
