+++ examples/transient_examples/PyFstat_example_short_transient_MCMC_search.py
@@ -1,4 +1,3 @@
-#!/usr/bin/env python
 """
 Short transient MCMC search
 ===========================
