+++ examples/mcmc_vs_grid_simple_example/PyFstat_example_simple_mcmc_vs_grid_comparison.py
@@ -1,4 +1,3 @@
-#!/usr/bin/env python
 """
 MCMC search v.s. grid search
 ============================
