### 2021-02-23 479f3e3
Fix up LineWriter's cache
- Fix some typos in the formatting
- Refactor the way in which cff is checked inside Writer to allow for
  LineWriter to benefit from it.
