### 2021-03-02 #3000 @khanx169
> 👋 @khanx169, @RobertRosca - It looks like you all have started, which is great. I just want to check in and see how things are going. Are you blocked on anything?

@danielskatz - thanks for checking in. The last two weeks were incredibly busy which caused a little bit of slowdown. However, I plan to pick up the pace again tomorrow and hopefully finish by this weekend if everything goes smoothly.
