### 2021-03-08 #3000 @khanx169
@danielskatz @dbkeitel I've gone through the review checklist and everything looks good to me. I've one minor suggestion: In the Github repo, there's a `run_all_examples.py` script. The docs point to GitHub, but it might be good to also explicitly mention this script there (especially since some examples need to be executed in order).
