### 2021-03-08 #3000 @dbkeitel
@khanx169 Thanks for the positive feedback!

> I've gone through the review checklist and everything looks good to me. I've one minor suggestion: In the Github repo, there's a run_all_examples.py script. The docs point to GitHub, but it might be good to also explicitly mention this script there (especially since some examples need to be executed in order).

This script was so far mostly intended for our own use - we run it as a manually-triggered action before releases to make sure the examples don't break. (But not automatically, because it takes much longer to run than the regular integration tests.) But if you found it useful, we're of course happy do document it more clearly.

To make sure, where are you suggesting to mention it?
- the README
- readthedocs
- both
