### 2021-03-08 #3000 @khanx169
> @khanx169 Thanks for the positive feedback!
>
> > I've gone through the review checklist and everything looks good to me. I've one minor suggestion: In the Github repo, there's a run_all_examples.py script. The docs point to GitHub, but it might be good to also explicitly mention this script there (especially since some examples need to be executed in order).
>
> This script was so far mostly intended for our own use - we run it as a manually-triggered action before releases to make sure the examples don't break. (But not automatically, because it takes much longer to run than the regular integration tests.) But if you found it useful, we're of course happy do document it more clearly.
>
> To make sure, where are you suggesting to mention it?
>
> * the README
> * readthedocs
> * both

@dbkeitel I was thinking specifically of the `Examples` section of readthedocs. I understand most users would probably want to look at some specific example at a time, but my thought process when I reached that section was: "It'd be nice if I could run all the examples first, and then come back and look at the scripts and outputs one by one to familiarize myself with the package". I suspect occasionally some odd user might think like this as well. Currently, the cumulative runtime for all examples is not terrible, but if more examples are added in the future, perhaps interactive notebooks (with the expected output for each cell already displayed) would be more useful.

I liked that the README is concise, but feel free to update both if you think that'd be more appropriate. Thanks.
