### 2021-03-08 #3000 @RobertRosca
Hey, I also went through the checklist, everything looks good!

I have some suggestions but none of them are major:

- Installation:
  - Pretty subjective, but I think having all the extra requirements under `extras_require` would be nice, then you could have some for `docs`, `style`, `test`, etc... which makes development and contribution a bit easier, also makes the CI scripts cleaner since everything can be a `pip install ".[style,test,...]"` command instead of using separate requirements files
  - The only objective benefit to this is that removing the requirements files makes it easier to maintain as you don't have to keep their pinned versions in sync
  - Also thanks for the quick response to my issue!
- Example usage:
  - Nice examples!
  - This PR https://github.com/PyFstat/PyFstat/pull/226 did work on automatically generating notebooks from the python scripts, I think using ipynb files directly would be nicer for the examples
  - That way when the examples get rendered in the docs they could include some rich output like plots
  - Same goes for the binder integration, currently the conversion from .py to .ipynb makes just one big cell with all the code in it, saving the notebook files directly (or using something like https://github.com/mwouts/jupytext) would let you have richer examples
  - And there could (where it makes sense) be some additional explanations about what's going on
  - This also makes it easier to include the examples as part of the tests directly with https://github.com/computationalmodelling/nbval
- Automated tests:
  - Codecov could be included as part of the CI, ran it locally and the coverage (with just the `tests.py` file) is around 62%
  - Also provides coverage details for PRs, letting you know if a contribution lowers the test coverage
  - Going back to the examples point, if the examples were ipynb files with the output saved in them then nbval could check that the output has not changed, and executing the examples would probably bump the coverage up a decent amount

The article mentions that one of the aims of PyFstat is to make this kind of analysis more accessible to new researchers and students, so I think converting the examples from python scripts to notebooks and then fleshing them out with more text and accompanying plots would be nice.

It's very easy to set up and run the examples yourself but I think that having the output saved and available in the docs makes a bit more beginner friendly. Reading through Asad Khan's suggestions we're thinking along the same lines :smile:

Overall this looks great! To be clear all the suggestions above are very minor, I'm happy with the current state of the submission and think it can be accepted now
