### 2021-03-11 PyFstat/PyFstat#287 @dbkeitel
document run_all_examples.py
suggestion from @khanx169 over on the JOSS review (https://github.com/openjournals/joss-reviews/issues/3000#issuecomment-792902860):

> I've gone through the review checklist and everything looks good to me. I've one minor suggestion: In the Github repo, there's a run_all_examples.py script. The docs point to GitHub, but it might be good to also explicitly mention this script there (especially since some examples need to be executed in order).

> I was thinking specifically of the Examples section of readthedocs. I understand most users would probably want to look at some specific example at a time, but my thought process when I reached that section was: "It'd be nice if I could run all the examples first, and then come back and look at the scripts and outputs one by one to familiarize myself with the package". I suspect occasionally some odd user might think like this as well. Currently, the cumulative runtime for all examples is not terrible, but if more examples are added in the future, perhaps interactive notebooks (with the expected output for each cell already displayed) would be more useful.

This should be quick to do. cc @Rodrigo-Tenorio
