### 2021-03-11 PyFstat/PyFstat#288 @dbkeitel
extra requirements management
from @RobertRosca over on the JOSS review (https://github.com/openjournals/joss-reviews/issues/3000#issuecomment-792999393):

> Pretty subjective, but I think having all the extra requirements under extras_require would be nice, then you could have some for docs, style, test, etc... which makes development and contribution a bit easier, also makes the CI scripts cleaner since everything can be a pip install ".[style,test,...]" command instead of using separate requirements files

> The only objective benefit to this is that removing the requirements files makes it easier to maintain as you don't have to keep their pinned versions in sync

I'll need to test a few permutations of this, but the suggestion sounds sensible.
