### 2021-03-11 PyFstat/PyFstat#289 @dbkeitel
revisit .py->.ipynb conversion for examples
from @RobertRosca over on the JOSS review (https://github.com/openjournals/joss-reviews/issues/3000#issuecomment-792999393):

> This PR PyFstat/PyFstat#226 did work on automatically generating notebooks from the python scripts, I think using ipynb files directly would be nicer for the examples
That way when the examples get rendered in the docs they could include some rich output like plots
Same goes for the binder integration, currently the conversion from .py to .ipynb makes just one big cell with all the code in it, saving the notebook files directly (or using something like https://github.com/mwouts/jupytext) would let you have richer examples
And there could (where it makes sense) be some additional explanations about what's going on
This also makes it easier to include the examples as part of the tests directly with https://github.com/computationalmodelling/nbval

We were definitely planning on revisiting this setup sometime soon, and some of these pointers will be very helpful to us. Thanks Robert!
