### 2021-03-11 PyFstat/PyFstat#290 @dbkeitel
code coverage
from @RobertRosca over on the JOSS review (https://github.com/openjournals/joss-reviews/issues/3000#issuecomment-792999393):

>  Codecov could be included as part of the CI, ran it locally and the coverage (with just the tests.py file) is around 62%
    Also provides coverage details for PRs, letting you know if a contribution lowers the test coverage
    Going back to the examples point, if the examples were ipynb files with the output saved in them then nbval could check that the output has not changed, and executing the examples would probably bump the coverage up a decent amount

Indeed we've been planning to do this for a while, I just never got around to reading the full setup documentation. But I know it's a very popular solution, so should be straightforward.
