### 2021-03-11 #3000 @dbkeitel
Thanks again to both reviewers! I've now opened issues over on our repo to track the suggestions.

I think we can resolve @khanx169's suggestion about the all-examples script quickly.

For @RobertRosca's suggestions, the examples handling and codecov are something we had planned (but not yet opened issues for) anyway, but the given pointers will be very helpful to it. The suggestion about installation / requirements management also sounds very sensible. It may however take us some time to get around to resolving these three points.
