### 2021-03-16 #3000 @dbkeitel
Hi @RobertRosca that would be awesome and much appreciated!
Generally, if there is no answer to one of the issues yet, nor a linked PR as in https://github.com/PyFstat/PyFstat/issues/287 , then you can safely assume we haven't started on it yet.
