### 2021-03-16 943a25e
Add more verbosity and how to do things
