### 2021-03-19 cff7a25
bump pillow to 8.1.2 in docs/
 - dependabot bumped it from 8.1.0 to 8.1.1 in https://github.com/PyFstat/PyFstat/pull/292
 - but next version is also a security fix: https://pillow.readthedocs.io/en/stable/releasenotes/8.1.2.html
