+++ docs/requirements.txt
@@ -2,4 +2,4 @@ sphinx==3.4.3
 sphinx_rtd_theme==0.5.1
 sphinx_gallery==0.8.2
 m2r2==0.2.7
-Pillow==8.1.1
+Pillow==8.1.2
