### 2021-03-26 #3000 @dbkeitel
@danielskatz I was wondering if there's a standard way to link to a package's ASCL entry in the left sidebar of the article, or if the easiest is to include it in the acknowledgments?
