### 2021-03-26 8a7fed5
Create codecov.yml
- starting from https://github.com/codecov/codecov-action#example-workflowyml-with-codecov-action
