### 2021-03-26 a7a187e
Update codecov.yml
- run `pytest tests.py`
