### 2021-03-26 c687e4f
Update codecov.yml
- add "install requirements" step
- add "Set up ephemerides" step
