+++ .github/workflows/codecov.yml
@@ -16,6 +16,19 @@ jobs:
       uses: actions/setup-python@master
       with:
         python-version: 3.8
+    - name: Install dependencies
+      run: |
+        python -m pip install --upgrade pip
+        pip install -r requirements.txt
+    - name: Set up ephemerides
+      run: |
+        export EPHEM_SRC=https://git.ligo.org/lscsoft/lalsuite/raw/master/lalpulsar/lib
+        echo "LALPULSAR_DATADIR: $LALPULSAR_DATADIR"
+        mkdir $LALPULSAR_DATADIR
+        wget $EPHEM_SRC/earth00-40-DE405.dat.gz -P $LALPULSAR_DATADIR
+        wget $EPHEM_SRC/sun00-40-DE405.dat.gz -P $LALPULSAR_DATADIR
+        echo "Successfully downloaded ephemerides:"
+        ls -l $LALPULSAR_DATADIR
     - name: Generate coverage report
       run: |
         pip install pytest
