+++ .github/workflows/run_examples.yml
@@ -22,7 +22,6 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         pip install -r requirements.txt
-        pip install git+https://gitlab.aei.uni-hannover.de/GregAshton/gridcorner
         pip install chainconsumer
     - name: Set up ephemerides
       run: |
