### 2021-03-26 637dc9c
Update codecov.yml
- add missing LALPULSAR_DATADIR env variable
