### 2021-03-26 f800c98
Update codecov.yml
- we seem to be generating only one file coverage.xml so only try to upload that one
- reduce test run to only TestWriter for the moment to speed up testing
