### 2021-03-26 09370bc
Update codecov.yml
- re-enable full tests.py run
