+++ .github/workflows/codecov.yml
@@ -34,7 +34,7 @@ jobs:
       run: |
         pip install pytest
         pip install pytest-cov
-        pytest tests.py::TestWriter --cov=./ --cov-report=xml
+        pytest tests.py --cov=./ --cov-report=xml
     - name: Upload coverage to Codecov
       uses: codecov/codecov-action@v1
       with:
