### 2021-03-26 PyFstat/PyFstat#290 @dbkeitel
@RobertRosca there's a basic working setup for this now in #296. Do you happen to know how we can exclude files from the calculation when using a github action setup? https://docs.codecov.io/docs/ignoring-paths applies to a setup with a top-level `codecov.yml` file but instead I went with a `.github/workflows/codecov.yml` action config, which while also yml seems to require a different structure.

* Should I just add that other file with the recommended syntax as well and the github action will know to get configuration options from it?
* Or would I do something like the following inside the action's own yml:
```
    - name: Generate coverage report
      with:
        ignore:
          - "setup.py"
          - "versioneer.py"
      run: |
        pip install pytest
        pip install pytest-cov
        pytest tests.py --cov=./ --cov-report=xml
```
