### 2021-03-26 #3000 @arfon
> @danielskatz I was wondering if there's a standard way to link to a package's ASCL entry in the left sidebar of the article, or if the easiest is to include it in the acknowledgments?

I'm afraid we can't support custom content in the left-hand column but you're welcome to add it as an explicit callout in the text (i.e., an acknowledgment) or cite directly as @danielskatz suggests.
