### 2021-03-26 #3000 @dbkeitel
Thanks both for checking, I'll just add it as suggested.

Also we'll be merging a few small infrastructure fixes/improvements soon (including codecov integration as suggested by @RobertRosca ), then make another release, and then would be ready for publication.
