### 2021-03-26 PyFstat/PyFstat#288 @Rodrigo-Tenorio
Following a commit in #285 , I should check whether it makes sense to have `black` and `flake8` in `pre-commit` *and* in `requirements-dev.txt` or only the former suffices.
