### 2021-03-26 4bd0945
JOSS paper: change ASCL reference to simple in-line link (looks better)
