### 2021-03-26 5343526
Revert "add PyFstat ASCL entry to biblio"
This reverts commit 8e437b37f474026dc6fc91ff3aaddbeddf945298.
