+++ joss-paper/paper.bib
@@ -249,17 +249,3 @@
     month = "12",
     year = "2020"
 }
-
-@misc{2021ascl.soft02027K,
-       author = {{Keitel}, David and {Tenorio}, Rodrigo and {Ashton}, Gregory and {Prix}, Reinhard},
-        title = "{PyFstat: Continuous gravitational-wave data analysis}",
-     keywords = {Software},
-         year = 2021,
-        month = feb,
-          eid = {ascl:2102.027},
-        pages = {ascl:2102.027},
-archivePrefix = {ascl},
-       eprint = {2102.027},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2021ascl.soft02027K},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
