### 2021-03-31 b9cded7
JOSS: grammar fix in acknowledgments
