### 2021-03-31 4d80492
JOSS: fix P.B. Covas biblio entries
