+++ joss-paper/paper.bib
@@ -188,7 +188,7 @@
 }
 
 @article{Covas:2020nwy,
-    author = "Covas, P.B. and Sintes, Alicia M.",
+    author = "Covas, P.~B. and Sintes, Alicia M.",
     title = "{First all-sky search for continuous gravitational-wave signals from unknown neutron stars in binary systems using Advanced LIGO data}",
     eprint = "2001.08411",
     archivePrefix = "arXiv",
@@ -202,7 +202,7 @@
 }
 
 @article{Covas:2020hcy,
-    author = "Covas, P. B.",
+    author = "Covas, P.~B.",
     title = "{Effects of proper motion of neutron stars on continuous gravitational-wave searches}",
     eprint = "2008.00983",
     archivePrefix = "arXiv",
