### 2021-03-31 #3000 @dbkeitel
@danielskatz : We've made a new [v1.11.4 release](https://github.com/PyFstat/PyFstat/releases/tag/v1.11.4) of the package including the following behind-the-scenes improvements, including reviewer comments:

- python 3.9 now supported
- improvements to documentation and examples
- for developers: flake8-docstrings and flake8-executable rules now enforced
- added a codemeta.json file
- started tracking test coverage with codecov

We'll continue improving documentation, examples and test coverage in the future, but for now we would be glad to proceed with publishing the paper. Two small final issues from our side, sorry for being difficult "customers" and thanks for the great experience so far:

1. I noticed a small inconsistency in citation formats:
 * papers with medium-length, explicitly given author lists become `et al.` in the compiled pdf, for example one with bibtex entry like `author = "Keitel, David and Prix, Reinhard and Papa, Maria Alessandra and Leaci, Paola and Siddiqi, Maham",`.
 * but for collaboration papers where inspire only puts e.g. `author = "Acernese, F. and others",`, these are displayed as `& others` in the pdf.
My (limited) understanding is that the bibtex+LaTeX combo normally knows that "and others" has to be translated to "et al.", but pandoc apparently doesn't. I tried a few things (just writing `et al.` in the bibtex author field -> disappears in the pdf; pasting full author lists (1000+ for some) -> works but compilation becomes extremely slow) but is there a recommended trick?

2. We'll also need to do a final update of our funding acknowledgments before "freezing" the paper.
