### 2021-04-01 03aefa5
add requirements.txt and tests.py to MANIFEST.in
 - hope this fixes conda-forge build #298
