+++ MANIFEST.in
@@ -1,4 +1,6 @@
 include versioneer.py
 include pyfstat/_version.py
 include CHANGELOG.md
+include requirements.txt
+include tests.py
 recursive-include examples/ *.py
