### 2021-04-01 #3000 @danielskatz
I'll see if I can find an answer to 1, and then we can do 2.  If we can't fix 1, I don't think it's a big problem though.
