### 2021-04-01 #3000 @arfon
> My (limited) understanding is that the bibtex+LaTeX combo normally knows that "and others" has to be translated to "et al.", but pandoc apparently doesn't. I tried a few things (just writing et al. in the bibtex author field -> disappears in the pdf; pasting full author lists (1000+ for some) -> works but compilation becomes extremely slow) but is there a recommended trick?

@tarleb - any suggestions here?
