+++ joss-paper/paper.bib
@@ -241,11 +241,14 @@
     author = "Abbott, R. and others",
     collaboration = "LIGO Scientific, Virgo",
     title = "{All-sky search in early O3 LIGO data for continuous gravitational-wave signals from unknown neutron stars in binary systems}",
-    journal = {arXiv e-prints},
     eprint = "2012.12128",
     archivePrefix = "arXiv",
     primaryClass = "gr-qc",
     reportNumber = "LIGO-P2000298",
-    month = "12",
-    year = "2020"
+    doi = "10.1103/PhysRevD.103.064017",
+    journal = "Phys. Rev. D",
+    volume = "103",
+    number = "6",
+    pages = "064017",
+    year = "2021"
 }
