### 2021-04-01 c81272d
JOSS: update funding acknowledgments
