### 2021-04-01 6b73ed0
JOSS: update Covas proper motion ref
