+++ joss-paper/paper.bib
@@ -202,16 +202,18 @@
 }
 
 @article{Covas:2020hcy,
-    author = "Covas, P.~B.",
+    author = "Covas, P. B.",
     title = "{Effects of proper motion of neutron stars on continuous gravitational-wave searches}",
     eprint = "2008.00983",
     archivePrefix = "arXiv",
     primaryClass = "astro-ph.IM",
     doi = "10.1093/mnras/staa3624",
-    month = "8",
+    journal = "Mon. Not. Roy. Astron. Soc.",
+    volume = "500",
+    number = "4",
+    pages = "5167--5176",
     year = "2020"
 }
-
 
 @article{Tenorio:2020cqm,
     author = "Tenorio, Rodrigo and Keitel, David and Sintes, Alicia M.",
