### 2021-04-01 b43d2a9
JOSS: consistent abbreviated MNRAS name
