+++ joss-paper/paper.bib
@@ -102,7 +102,7 @@
 @article{Vousden:2015pte,
     author = {Vousden, W. D. and Farr, W. M. and Mandel, I.},
     title = "{Dynamic temperature selection for parallel tempering in Markov chain Monte Carlo simulations}",
-    journal = {Monthly Notices of the Royal Astronomical Society},
+    journal = "Mon. Not. Roy. Astron. Soc.",
     volume = {455},
     number = {2},
     pages = {1919-1937},
@@ -110,8 +110,8 @@
     month = {11},
     issn = {0035-8711},
     archivePrefix = {arXiv},
-       eprint = {1501.05823},
- primaryClass = {astro-ph.IM},
+    eprint = {1501.05823},
+    primaryClass = {astro-ph.IM},
     doi = {10.1093/mnras/stv2422},
     url = {https://doi.org/10.1093/mnras/stv2422},
 }
