### 2021-04-01 ab2639f
JOSS: try verbatim "et al." in online generator, seems to behave slightly differently than local Make
