+++ joss-paper/paper.bib
@@ -86,7 +86,7 @@
 }
 
 @article{TheLIGOScientific:2014jea,
-    author = "Aasi, J. and others",
+    author = "Aasi, J. et al.",
     collaboration = "LIGO Scientific",
     title = "{Advanced LIGO}",
     eprint = "1411.4547",
