### 2021-04-01 cd4f779
JOSS: just use 3 first authors + "and others" for collaboration papers to get "et al."
