+++ joss-paper/paper.bib
@@ -71,7 +71,7 @@
 }
 
 @article{TheVirgo:2014hva,
-    author = "Acernese, F. and others",
+    author = "{Acernese}, F. and {Agathos}, M. and {Agatsuma}, K. and others",
     collaboration = "VIRGO",
     title = "{Advanced Virgo: a second-generation interferometric gravitational wave detector}",
     eprint = "1408.3978",
@@ -86,7 +86,7 @@
 }
 
 @article{TheLIGOScientific:2014jea,
-    author = "Aasi, J. et al.",
+    author = "{Aasi}, J. and {Abbott}, B.~P. and {Abbott}, R. and others",
     collaboration = "LIGO Scientific",
     title = "{Advanced LIGO}",
     eprint = "1411.4547",
@@ -240,7 +240,7 @@
 }
 
 @article{Abbott:2020mev,
-    author = "Abbott, R. and others",
+    author = "{Abbott}, R. and {Abbott}, T.~D. and {Abraham}, S.  and others",
     collaboration = "LIGO Scientific, Virgo",
     title = "{All-sky search in early O3 LIGO data for continuous gravitational-wave signals from unknown neutron stars in binary systems}",
     eprint = "2012.12128",
