### 2021-04-01 #3000 @dbkeitel
- citation formats: Now I've simply resorted to entering the first 3 authors + "and others" in the `.bib` file, the result is the correct "first tauthor et al." in the pdf text. Of course the bibliography entries at the end themselves now have those 3 + "& others"; if that's ok with JOSS style, this solution is good for me.
- acknowledgments: also updated
- deposited version: we're automatically depositing our releases on Zenodo, but for the one I did yesterday there has been an issue with the PyPI/Conda uploads, which I'll try to fix and then make another proper "reference version" for the paper, and will post the corresponding Zenodo DOI here.
