### 2021-04-01 PyFstat/PyFstat#290 @RobertRosca
Hi, sorry for the late reply, I've been very busy recently and didn't have much time to look through my open source stuff.

I noticed that the PR is already merged, did you end up solving this issue? Part of the problem is that you do `cov=./`, which will try to calculate the coverage for all the files in the repository. If you instead do `cov=pyfstat` then it will only compute coverage for the files that are in the module itself (e.g. anything that is imported when you do `import pyfstat...`, so not `setup.py` or `versioneer.py`).
