### 2021-04-01 PyFstat/PyFstat#290 @dbkeitel
Thanks @RobertRosca just changing the top level is a good idea! I'll make another PR for that. The reports also seem more useful if there is already a base version on master, so it's good to do this incrementally.
