### 2021-04-01 0d65c33
codecov.yml: only run on pyfstat package dir
- as suggest by @RobertRosca in #290
