### 2021-04-01 fa294ea
codecov: ignore pyfstat/_version.py
