+++ .github/codecov.yml
@@ -18,3 +18,6 @@ comment:
   layout: "reach,diff,flags,files,footer"
   behavior: default
   require_changes: no
+  
+ignore:
+  - "pyfstat/_version.py"
