### 2021-04-01 PyFstat/PyFstat#290 @dbkeitel
For ignoring the remaining auto-generated `pyfstat/_version.py` file, I've now added a `.github/codecov.yml` file (re-reading the docs, I now understand the difference between the config and workflow files, it's just confusing that they're both .yml) and in #301 I try to use it for that ignore feature. But somehow all actions fail with an "internal error" at the moment.
