### 2021-04-02 e1c60e7
README: add codecov badge, sort badges
