### 2021-04-02 PyFstat/PyFstat#290 @dbkeitel
Had to move from `.github/codecov.yml` to `codecov.yml` because any non-root-directory location seems to be only semi-supported. (#302)

I think this now works as it should, and we can start looking at actually increasing coverage. (Maybe fine-tune the config a bit more, ignoring "paranoia exceptions" and things like that.) Thanks a lot @RobertRosca for pushing me in the right direction!
