### 2021-04-02 #3000 @dbkeitel
@danielskatz v1.11.5 of the package, with DOI [10.5281/zenodo.4659766](https://doi.org/10.5281/zenodo.4659766), is ready as a reference version for the paper.

However, it doesn't have the longer title of the paper, and the order of the author list is different (original author first). Should I make a separate manual deposit off the `joss-paper` branch with exactly matching metadata, or is this one fine?
