### 2021-04-02 #3000 @dbkeitel
> @dbkeitel - you should be able to go to the Zenodo page and edit the metadata

Of course, but then it would look odd that this one version has different data than all others before and after. If JOSS needs one deposit that exactly matches the paper, then I'd prefer to do a "duplicate", if that's ok.
