### 2021-04-02 #3000 @danielskatz
> Of course, but then it would look odd that this one version has different data than all others before and after.

As Emerson said, "A foolish consistency is the hobgoblin of little minds" - but it's up to you :)

> If JOSS needs one deposit that exactly matches the paper, then I'd prefer to do a "duplicate", if that's ok.

Sure, that's fine.
