### 2021-04-05 df3b022
minor changes to paper
