### 2021-04-05 #3000 @danielskatz
My proof read has found some minor changes that are needed - see https://github.com/PyFstat/PyFstat/pull/304.  Please merge this, or let me know what you disagree with.
