+++ joss-paper/paper.md
@@ -150,7 +150,7 @@ the Spanish Ministerio de Ciencia, Innovación y Universidades and Agencia Estat
 PID2019-106416GB-I00/AEI/10.13039/501100011033,
 RED2018-102661-T,
 RED2018-102573-E,
-FPA2017-90687-REDC
+FPA2017-90687-REDC,
 FPU 18/00694,
 and BEAGAL 18/00148 (cofinanced by the Universitat de les Illes Balears);
 the Comunitat Autonoma de les Illes Balears
